Improve the proposed algorithm
This improves the proposed algorithm to achieve origin isolation in several scenarios that were previously prevented due to existing site-keyed agent clusters. It adds a scenarios.md document detailing various tricky scenarios and how the algorithm responds. This shows that the algorithm is explicitly racy in some cases (as was the previous algorithm); we believe this to be unavoidable.

The exact details of the algorithm are still under some discussion, in particular the check as to whether there exist same-origin realms in a site-keyed agent cluster. See #6 for that.